[REF] Move Arquivo Certificado class to a separated file
